Renamed drinkAoi -> APi.
