+++ app/src/main/java/com/tkachuk/cocktailbar/injection/module/NetworkModule.kt
@@ -1,6 +1,6 @@
 package com.tkachuk.cocktailbar.injection.module
 
-import com.tkachuk.cocktailbar.network.DrinkApi
+import com.tkachuk.cocktailbar.network.Api
 import com.tkachuk.cocktailbar.util.BASE_URL
 import dagger.Module
 import dagger.Provides
@@ -24,8 +24,8 @@ object NetworkModule {
     @Provides
     @Reusable
     @JvmStatic
-    internal fun provideDrinkApi(retrofit: Retrofit): DrinkApi {
-        return retrofit.create(DrinkApi::class.java)
+    internal fun provideDrinkApi(retrofit: Retrofit): Api {
+        return retrofit.create(Api::class.java)
     }
 
     /**
