+++ app/src/main/java/com/tkachuk/cocktailbar/network/Api.kt
@@ -7,7 +7,7 @@ import io.reactivex.Observable
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-interface DrinkApi {
+interface Api {
 
     //List of ingredients
     @GET("list.php?i=list")
